Moving common code to Base strategy and re-using
